Rename FigCommon to common
