+++ lineage/figures/figure11.py
@@ -2,8 +2,7 @@
 import pickle
 from string import ascii_lowercase
 
-from .figureCommon import getSetup, plot_all
-from ..plotTree import plot_networkx
+from .common import getSetup, plot_all
 
 concs = ["Control", "Lapatinib 25 nM", "Lapatinib 50 nM", "Lapatinib 250 nM"]
 concsValues = ["Control", "25 nM", "50 nM", "250 nM"]
