+++ lineage/figures/figure14.py
@@ -4,7 +4,7 @@ from networkx.drawing.nx_agraph import to_agraph
 
 from .figure11 import lapt_tHMMobj_list
 from .figure12 import gemc_tHMMobj_list
-from .figureCommon import getSetup, subplotLabel
+from .common import getSetup, subplotLabel
 
 gemc = gemc_tHMMobj_list[0]
 lapt = lapt_tHMMobj_list[0]
